Move images in assets folder to its own folder.
