+++ src/assets/images/repo-icon.png
