+++ src/components/UserCard/UserCardBack/UserCardBack.css
@@ -4,7 +4,7 @@
   padding: 0rem 0.5rem;
 
   overflow: hidden;
-  background: url(./../../../assets/swirl-background.jpg);
+  background: url(./../../../assets/images/swirl-background.jpg);
   background-size: 20rem;
   border: solid #00025f 1rem;
   border-radius: 1rem;
